Partial veneer changes for vision
